git-svn-id: https://stlport.svn.sourceforge.net/svnroot/stlport/trunk/STLport@2712 01b27279-081b-0410-8cf9-d9d9e0fb0389

+++ test/unit/unordered_test.cpp
@@ -462,7 +462,7 @@ void UnorderedTest::equal_range()
 struct Key
 {
   Key() : m_data(0) {}
-  explicit Key(size_t data) : m_data(data) {}
+  explicit Key(int data) : m_data(data) {}
 
   int m_data;
 };
